Fix for f5_snatpool type/provider
The f5_snatpool type and provider were able to display the current
state of a snatpool, and were even able to make changes, but still
displayed a message that they were changing resources even when
they weren't.

This change allows for comparison of structured data by reducing it
to a string and comparing it that way.  It's not the most ideal way
to do it, but it's the method we currently have at our disposal.