Adding a first pass at f5_profileclientssl
I tested it with the following code:

f5_profileclientssl { "demo2":
  ensure                  => present,
  certificate_file        => {"value" => "demo2-server.crt", "default_flag" => false},
  key_file                => {"value" => "demo2-server.key", "default_flag" => false},
  ca_file                 => {"value" => "demo2-ca.crt", "default_flag" => false},
  peer_certification_mode => {"value" => "CLIENTSSL_CERTIFICATE_MODE_AUTO", "default_flag" => false},
}

I tried to stick as close to the iControl API as possible.  In doing so I have
made it so that the values of the parameters take the ProfileString.  This
allows for easy comparison since the data structure returned from the icontrol
gem also works with hashes.