(#10849) Fix f5 facts to use string instead of sym.
Facter need the hash to use string value keys instead of symbol keys
otherwise the values does not appear to be accessible in puppet
manifests. This patch changes facts[:key] to facts['key'] so $key is
available in puppet.

This patch also adds partition to the available facts since we support
partition in F5.