(#12000) Update f5_virtualserver type for consistency.
F5_virtualserver is converting profile_persistence values to
"true"/"false", so we change it to true/false for consistency across f5
types.

F5_virtualserver appears to only toggle profile_persistence settings to
default_profile => false, so the documentation have been updated. It's
unknown if we can configure the value to true in combination with other
settings, but at this moment it does not appear possible via the web UI
interface well.